Changes based on review comments - revert code added to switch to old Prison API endpoint.
