+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/client/PrisonApiClient.kt
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.prison.api.OffenderNonAssociationDetailsDto
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.prison.api.PrisonerHousingLocationsDto
-import uk.gov.justice.digital.hmpps.visitscheduler.dto.prison.api.PrisonerUnitCodeDto
 import java.time.Duration
 
 @Component
@@ -17,8 +16,6 @@ class PrisonApiClient(
 ) {
   private val offenderNonAssociationDetails = object : ParameterizedTypeReference<OffenderNonAssociationDetailsDto>() {}
 
-  private val prisonerDetail = object : ParameterizedTypeReference<PrisonerUnitCodeDto>() {}
-
   private val prisonerHousingLocationsDto = object : ParameterizedTypeReference<PrisonerHousingLocationsDto>() {}
 
   fun getOffenderNonAssociation(offenderNo: String): OffenderNonAssociationDetailsDto? {
@@ -26,14 +23,6 @@ class PrisonApiClient(
       .uri("/api/offenders/$offenderNo/non-association-details")
       .retrieve()
       .bodyToMono(offenderNonAssociationDetails)
-      .block(apiTimeout)
-  }
-
-  fun getPrisonerDetails(offenderNo: String): PrisonerUnitCodeDto? {
-    return webClient.get()
-      .uri("/api/prisoners/$offenderNo/full-status")
-      .retrieve()
-      .bodyToMono(prisonerDetail)
       .block(apiTimeout)
   }
