--- src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/dto/prison/api/PrisonApiGetLevelsEndpoint.kt
@@ -1,6 +1,0 @@
-package uk.gov.justice.digital.hmpps.visitscheduler.dto.prison.api
-
-enum class PrisonApiGetLevelsEndpoint {
-  FULL_STATUS,
-  HOUSING_LOCATION;
-}
