--- src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/dto/prison/api/PrisonerDetailsDto.kt
@@ -1,5 +1,0 @@
-package uk.gov.justice.digital.hmpps.visitscheduler.dto.prison.api
-
-interface PrisonerDetailsDto {
-  fun getLevels(): Map<PrisonerHousingLevels, String?>
-}
